# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "FilesWithTag API",
        "title": "FilesWithTag API",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8050",
    "basePath": "/",
    "paths": {
...
        "/Settings": {
            "get": {
                "description": "Получение настроек запуска",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "Получение настроек",
                "operationId": "getSettings",
                "responses": {
                    "200": {
                        "description": "settings",
                        "schema": {
                            "$ref": "#/definitions/settingmodule.Setting"
                        }
                    }
                }
            },
            "put": {
                "description": "Получение настроек запуска",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "Получение настроек",
                "operationId": "getSettings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name setting",
                        "name": "NameSetting",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name setting",
                        "name": "ValueSetting",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "settings",
                        "schema": {
                            "$ref": "#/definitions/service.Responce"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/service.ResponceError"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/service.ResponceError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "service.Responce": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "service.ResponceError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "settingmodule.Setting": {
            "type": "object",
            "properties": {
                "BasePath": {
                    "type": "string"
                },
                "LastDir": {
                    "type": "string"
                },
                "Port": {
                    "type": "string"
                }
            }
        }
    }
}